@@ -1,58 +1,48 @@
 {
-  "name": "webui",
+  "name": "client-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "build:ssr": "ng run WebUI:server:dev",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "8.2.12",
-    "@angular/common": "8.2.12",
-    "@angular/compiler": "8.2.12",
-    "@angular/core": "8.2.12",
-    "@angular/forms": "8.2.12",
-    "@angular/platform-browser": "8.2.12",
-    "@angular/platform-browser-dynamic": "8.2.12",
-    "@angular/platform-server": "8.2.12",
-    "@angular/router": "8.2.12",
-    "@nguniversal/module-map-ngfactory-loader": "8.1.1",
-    "bootstrap": "^4.6.0",
-    "core-js": "^3.8.3",
-    "jquery": "^3.5.1",
-    "oidc-client": "^1.11.3",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.6.3",
-    "zone.js": "0.9.1"
+    "@angular/animations": "~12.0.0",
+    "@angular/common": "~12.0.0",
+    "@angular/compiler": "~12.0.0",
+    "@angular/core": "~12.0.0",
+    "@angular/forms": "~12.0.0",
+    "@angular/platform-browser": "~12.0.0",
+    "@angular/platform-browser-dynamic": "~12.0.0",
+    "@angular/router": "~12.0.0",
+    "@fortawesome/angular-fontawesome": "^0.9.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.35",
+    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "bootstrap": "^5.0.1",
+    "ngx-order-pipe": "^2.1.1",
+    "ngx-pagination": "^5.1.0",
+    "oidc-client": "^1.11.5",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.1.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@types/jasmine": "~3.4.4",
-    "@types/jasminewd2": "~2.0.8",
-    "@types/node": "~12.11.6",
-    "codelyzer": "^5.2.2",
-    "ini": "^1.3.7",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^5.2.3",
+    "@angular-devkit/build-angular": "~12.0.0",
+    "@angular/cli": "~12.0.0",
+    "@angular/compiler-cli": "~12.0.0",
+    "@types/datatables.net": "^1.10.18",
+    "@types/jasmine": "~3.6.0",
+    "@types/jquery": "^3.3.33",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~3.7.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.5.4",
-    "typescript": "3.5.3"
-  },
-  "optionalDependencies": {
-    "node-sass": "^5.0.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.4.1",
-    "tslint": "~5.20.0"
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "typescript": "~4.2.3"
   }
 }
