@@ -2,23 +2,27 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { AuthenticationRoutingModule } from './authentication-routing.module';
-import { LoginComponent } from './components/login/login.component';
 import { LoginMenuComponent } from './components/login-menu/login-menu.component';
-import { LogoutComponent } from './components/logout/logout.component';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+import { LoginComponent } from './components/login/login.component';
+import { LogoutComponent } from './components/logout/logout.component';
 
 
 @NgModule({
   declarations: [
-    LoginComponent,
     LoginMenuComponent,
+    LoginComponent,
     LogoutComponent
   ],
   imports: [
     CommonModule,
     AuthenticationRoutingModule,
+    NgbModule,
     FontAwesomeModule
   ],
-  exports: [LoginMenuComponent, LoginComponent, LogoutComponent]
+  exports:[
+    LoginMenuComponent
+  ]
 })
 export class AuthenticationModule { }
