@@ -3,16 +3,19 @@ import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './header/header.component';
 import { FooterComponent } from './footer/footer.component';
 import { RouterModule } from '@angular/router';
-import { ApiAuthorizationModule } from 'src/api-authorization/api-authorization.module';
-
-
+import { AuthenticationModule } from '../modules/authentication/authentication.module';
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
-  declarations: [HeaderComponent, FooterComponent],
+  declarations: [
+    HeaderComponent,
+    FooterComponent
+  ],
   imports: [
     CommonModule,
     RouterModule,
-    ApiAuthorizationModule
+    AuthenticationModule,
+    HttpClientModule
   ],
   exports:[
     HeaderComponent,
