@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { SeriesService } from './series.service';
 
 describe('SeriesService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: SeriesService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SeriesService);
+  });
 
   it('should be created', () => {
-    const service: SeriesService = TestBed.get(SeriesService);
     expect(service).toBeTruthy();
   });
 });
