@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { UniversesService } from './universes.service';
 
 describe('UniversesService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: UniversesService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(UniversesService);
+  });
 
   it('should be created', () => {
-    const service: UniversesService = TestBed.get(UniversesService);
     expect(service).toBeTruthy();
   });
 });
