@@ -2,13 +2,13 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   templateUrl: './grade.component.html',
-  styleUrls: ['./grade.component.css']
+  styleUrls: ['./grade.component.scss']
 })
 export class GradeComponent implements OnInit {
 
   constructor() { }
 
-  ngOnInit() {
+  ngOnInit(): void {
   }
 
 }
