@@ -2,13 +2,13 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   templateUrl: './manufacturer.component.html',
-  styleUrls: ['./manufacturer.component.css']
+  styleUrls: ['./manufacturer.component.scss']
 })
 export class ManufacturerComponent implements OnInit {
 
   constructor() { }
 
-  ngOnInit() {
+  ngOnInit(): void {
   }
 
 }
