@@ -2,13 +2,13 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   templateUrl: './scale.component.html',
-  styleUrls: ['./scale.component.css']
+  styleUrls: ['./scale.component.scss']
 })
 export class ScaleComponent implements OnInit {
 
   constructor() { }
 
-  ngOnInit() {
+  ngOnInit(): void {
   }
 
 }
