@@ -1,10 +1,9 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './pages/home/home.component';
 
-
 const routes: Routes = [
-  {path:'',component:HomeComponent}
+  {path: '', component:HomeComponent}
 ];
 
 @NgModule({
